+++ CHANGES.md
@@ -4,8 +4,11 @@
 
 *Release to be disclosed soon*
 
+* Changed `to_serialize` in `JSONAPI.serialize` from expecting a list or
+  collection to also expecting a query.
+
 ## Version 0.1
 
-*Release to be disclosed soon*
+*2014-07-20*
 
 * Added `JSONAPI`, `JSONAPIMixin`, and `as_relationship`